+++ tests/utils/sc_linac/test_cavity.py
@@ -42,7 +42,7 @@ from sc_linac_physics.utils.sc_linac.linac_utils import (
     ALL_CRYOMODULES,
 )
 from sc_linac_physics.utils.sc_linac.piezo import Piezo
-from tests.utils.mock_utils import mock_func
+from tests.mock_utils import mock_func
 
 
 @pytest.fixture
