+++ tests/utils/sc_linac/test_stepper.py
@@ -13,7 +13,7 @@ from sc_linac_physics.utils.sc_linac.linac_utils import (
     ALL_CRYOMODULES,
 )
 from sc_linac_physics.utils.sc_linac.stepper import StepperTuner
-from tests.utils.mock_utils import mock_func
+from tests.mock_utils import mock_func
 
 
 @pytest.fixture
